metamcp raw sse
